+++ apps/web/src/types/tournament.ts
@@ -46,3 +46,13 @@ export type PostCreateTournamentRequestT = {
 export type PostCreateTournamentResponseT = {
   tournamentId: number;
 };
+
+/** 초대 코드 / 토너먼트 미리보기 응답 */
+export type GetInvitePreviewResponseT = {
+  tournamentId: number;
+  tournamentName: string;
+  itemCount: number;
+  participantCount: number;
+  /** 요청자(게스트/회원)가 이미 이 토너먼트에 참여 중인지. 미인증·미참여면 false */
+  joined: boolean;
+};
